requirements.txt
alembic==1.6.5
click==8.0.1
Flask==2.0.1
Flask-Migrate==3.0.1
flask-serialize==1.5.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
permissive-dict==1.0.4
psycopg2-binary==2.9.1
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.18.0
python-editor==1.0.4
python-statemachine==0.8.0
six==1.16.0
SQLAlchemy==1.4.21
SQLAlchemy-serializer==1.4.1
Werkzeug==2.0.1
